financial_calculaor.py

+++ functions/notes.C
@@ -2,10 +2,10 @@
 #include <stdio.h>
 int num;
 char name[50],verb[50];
-float add(int numOne, int numTwo){
+int add(int numOne, int numTwo){
     return numOne + numTwo;
 }
-float due(char assignment[50], char day [20]){
+int due(char assignment[50], char day [20]){
     printf("The %s assignment is due %s", assignment, day)
 }
